+++ ricaun.RevitTest.TestAdapter/Services/RevitTestConsole.cs
@@ -32,10 +32,11 @@ namespace ricaun.RevitTest.TestAdapter.Services
             if (string.IsNullOrWhiteSpace(applicationPath))
                 return null;
 
-            AdapterLogger.Logger.Info($"Application: {Path.GetFileName(applicationPath)}", 0);
+            AdapterLogger.Logger.InfoAny($"Application: {Path.GetFileName(applicationPath)}");
 
             applicationPath = GetEnvironmentVariable(applicationPath);
 
+            AdapterLogger.Logger.DebugOnlyLocal($"Application Environment: {applicationPath}");
             if (ApplicationUtils.Download(applicationPath, out string directory))
             {
                 AdapterLogger.Logger.Info($"Application Download: {Path.GetFileName(applicationPath)}");
@@ -53,6 +54,7 @@ namespace ricaun.RevitTest.TestAdapter.Services
                 }
             }
 
+            AdapterLogger.Logger.Info($"Application Download: Fail");
             return null;
         }